9. chatbot & page css_1

+++ src/main/java/com/project/joonggo/domain/ChatRoomVO.java
@@ -9,7 +9,7 @@ import lombok.*;
 @NoArgsConstructor
 public class ChatRoomVO {
     private int roomId;
-    private int userNum;
+    private long userNum;
     private String roomName;
     private String roomStatus;
 }
